pyproject.toml
[project]
name = "langchain-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.13",
    "langchain-core>=1.2.21",
    "langchain-openai>=1.1.12",
    "python-dotenv>=1.2.2",
]
